Add comprehensive profiling system with Spall integration
